refactor(animation): adjust animation type definition
