+++ packages/g6/src/animations/types.ts
@@ -1,22 +1,54 @@
-import type { IAnimation, IAnimationEffectTiming } from '@antv/g';
-import type { Element, State } from '../types';
+import type { IAnimation } from '@antv/g';
+import type { AnimationStage } from '../spec/element/animation';
+import type { Element, ElementType, State } from '../types';
+
+export type STDAnimation = AnimationOptions[];
 
 /**
- * <zh/> 为 string 时，会从注册的动画中获取
+ * <zh/> 元素动画选项
  *
- * <en/> When it is a string, it will be obtained from the registered animation
+ * <en/> Element animation options
  */
-export type Animation = false | string | STDAnimation;
-
-export type STDAnimation = ConfigurableAnimationOptions[];
-
-interface ConfigurableAnimationOptions extends AnimationEffectTiming {
+export interface AnimationOptions extends AnimationEffectTiming {
+  /**
+   * <zh/> 执行动画的字段（样式）名
+   *
+   * <en/> Field (style) name of the animation
+   */
   fields: string[];
+  /**
+   * <zh/> 执行动画的图形，默认为当前元素
+   *
+   * <en/> Shape of the animation, default is the current element
+   */
   shape?: string;
+  /**
+   * <zh/> 参与动画的状态
+   *
+   * <en/> States involved in the animation
+   */
   states?: State[];
 }
 
 export interface AnimationContext {
+  /**
+   * <zh/> 执行动画的元素
+   *
+   * <en/> Element to execute animation
+   */
+  element: Element;
+  /**
+   * <zh/> 元素类型
+   *
+   * <en/> Element type
+   */
+  elementType: ElementType;
+  /**
+   * <zh/> 动画阶段
+   *
+   * <en/> Animation stage
+   */
+  stage: AnimationStage;
   /**
    * <zh/> 动画的源样式
    *
@@ -36,32 +68,61 @@ export interface AnimationContext {
    *
    * <en/> For example, before the element is destroyed, the final state opacity of the element needs to be set to 0
    */
-  modifiedStyle?: Record<string, unknown>;
+  modifiedStyle: Record<string, unknown>;
   /**
-   * <zh/> 动画配置过滤
+   * <zh/> 动画时序配置
    *
-   * <en/> Animation configuration filter
-   * @remarks
-   * <zh/> 某些情况下，需要过滤掉一些动画配置，例如位移动画只会作用于元素根节点图形，不会作用于子图形。这项配置通常用于优化动画性能，避免不必要的动画执行
-   *
-   * <en/> In some cases, it is necessary to filter out some animation configurations. For example, the displacement animation will only affect the root node shape of the element, not the child shape. This configuration is usually used to optimize animation performance and avoid unnecessary animation execution
+   * <en/> Animation effect timing
    */
-  animationsFilter?: (animation: ConfigurableAnimationOptions) => boolean;
-  /**
-   * <zh/> 元素状态
-   *
-   * <en/> Element states
-   */
-  states: State[];
+  // effectTiming: AnimationEffectTiming;
 }
 
-export type AnimationEffectTiming = Partial<
-  Pick<IAnimationEffectTiming, 'duration' | 'delay' | 'easing' | 'iterations' | 'direction' | 'fill'>
->;
+/**
+ * <zh/> 动画效果时序
+ *
+ * <en/> Animation effect timing
+ */
+export interface AnimationEffectTiming {
+  /**
+   * <zh/> 动画延迟时间
+   *
+   * <en/> Animation delay time
+   */
+  delay?: number;
+  /**
+   * <zh/> 动画方向
+   *
+   * <en/> Animation direction
+   */
+  direction?: PlaybackDirection;
+  /**
+   * <zh/> 动画持续时间
+   *
+   * <en/> Animation duration
+   */
+  duration?: number;
+  /**
+   * <zh/> 动画缓动函数
+   *
+   * <en/> Animation easing function
+   */
+  easing?: string;
+  /**
+   * <zh/> 动画结束后的填充模式
+   *
+   * <en/> Fill mode after the animation ends
+   */
+  fill?: FillMode;
+  /**
+   * <zh/> 动画迭代次数
+   *
+   * <en/> Number of iterations of the animation
+   */
+  iterations?: number;
+}
 
 export type AnimationExecutor = (
-  shape: Element,
-  animation: Animation | false,
-  baseEffectTiming: AnimationEffectTiming,
-  context: AnimationContext,
+  element: Element,
+  keyframes: [Record<string, unknown>, Record<string, unknown>],
+  options: STDAnimation,
 ) => IAnimation | null;
