+++ packages/g6/src/spec/element/combo.ts
@@ -1,7 +1,8 @@
+import type { AnimationOptions } from '../../animations/types';
 import type { BaseComboStyleProps } from '../../elements/combos';
 import type { CallableObject } from '../../types';
 import type { ComboData } from '../data';
-import type { AnimationOptions } from './animation';
+import type { AnimationStage } from './animation';
 import type { PaletteOptions } from './palette';
 
 /**
@@ -33,7 +34,7 @@ export interface ComboOptions {
    *
    * <en/> Combo animation
    */
-  animation?: AnimationOptions;
+  animation?: false | Record<AnimationStage, false | string | AnimationOptions[]>;
   /**
    * <zh/> 色板
    *
@@ -45,7 +46,7 @@ export interface ComboOptions {
 export interface StaticComboOptions {
   style?: ComboStyle;
   state?: Record<string, ComboStyle>;
-  animation?: AnimationOptions;
+  animation?: false | Record<AnimationStage, false | string | AnimationOptions[]>;
   palette?: PaletteOptions;
 }
